+++ src/frontend/apps/drive/src/features/explorer/components/filters/index.ts
@@ -1,6 +1,7 @@
 export { ExplorerFilters } from "./ExplorerFilters";
-export { handleFilterChange } from "./filterUtils";
+export { ALL, handleFilterChange } from "./filterUtils";
 export { ExplorerFilterContact } from "./ExplorerFilterContact";
+export { ExplorerFilterLocation } from "./ExplorerFilterLocation";
 export { ExplorerFilterCategory } from "./ExplorerFilterCategory";
 export { ExplorerFilterModified } from "./ExplorerFilterModified";
 export { ExplorerFilterType } from "./ExplorerFilterType";